modifies products-API-Coontroller and creates folders for api controllers and routes
