+++ website/src/controllers/api/users-API-controller.js
@@ -1,0 +1,10 @@
+const userService = require("../../services/userService");
+
+module.exports = {
+  ApiUsers: (req, res) => {
+    // metodos necesarios
+  },
+  APiUserDetail: (req, res) => {
+    // metodos necesarios
+  },
+};
